refactor(frontend): extract processing audit tab
