+++ frontend/src/pages/DatasetAudit.tsx
@@ -18,8 +18,6 @@ import {
 	Collapse,
 	Badge,
 	DatePicker,
-	Drawer,
-	Alert,
 } from "antd";
 import type { ColumnsType } from "antd/es/table";
 import {
@@ -34,6 +32,7 @@ import { useDatasetById } from "../hooks/useDatasets";
 import { useAuditDatasets } from "../hooks/useAuditDatasets";
 import type { AuditDataset } from "../hooks/useAuditDatasets";
 import DatasetAuditDetail from "../components/DatasetAudit/DatasetAuditDetail";
+import ProcessingAuditTab, { PROCESSING_ACTIVE_STATUSES } from "../components/DatasetAudit/ProcessingAuditTab";
 import { useDatasetAudits, DatasetAuditUserInfo, useDatasetContributors } from "../hooks/useDatasetAudit";
 import { supabase } from "../hooks/useSupabase";
 import { useFlaggedDatasets } from "../hooks/useDatasetFlags";
@@ -42,7 +41,7 @@ import { useAuditNavigation } from "../hooks/useAuditNavigation";
 import { usePendingCorrections } from "../hooks/usePendingCorrections";
 import { palette } from "../theme/palette";
 import { getBiomeEmoji, getBiomeTagColor, truncateBiomeLabel } from "../utils/biomeDisplay";
-import { useDatasetLogs, useProcessingOverview, ProcessingOverviewRow, ProcessingStatus } from "../hooks/useProcessingOverview";
+import { useProcessingOverview } from "../hooks/useProcessingOverview";
 import { useIsMobile } from "../hooks/useIsMobile";
 import { useAnalytics } from "../hooks/useAnalytics";
 import { isDatasetReadyForAudit } from "../utils/processingSteps";
@@ -78,13 +77,6 @@ const PROCESSING_STATE_TO_FIELD: Record<ProcessingStateFilterKey, keyof AuditDat
 const DEFAULT_PROCESSING_STATE_FILTERS: ProcessingStateFilterKey[] = PROCESSING_STATE_OPTIONS.map((option) => option.value);
 
 const BADGE_OVERFLOW_COUNT = 999999;
-const DEFAULT_PROCESSING_STATUS_FILTERS: ProcessingStatus[] = ["QUEUED", "PROCESSING", "FAILED"];
-
-const formatHours = (value: number | null | undefined): string => {
-	if (typeof value !== "number" || Number.isNaN(value)) return "—";
-	return `${value.toFixed(1)}h`;
-};
-
 // Helper function to check if a dataset is ready to be audited.
 // Auditing only depends on the legacy pipeline; it must not be gated on the v2
 // combined segmentation model, AOI generation, or the current processing status.
@@ -242,14 +234,6 @@ function DatasetAuditInner() {
 	const [hasProcessingStates, setHasProcessingStates] = useState<ProcessingStateFilterKey[]>(DEFAULT_PROCESSING_STATE_FILTERS);
 	const [inSeasonOnly, setInSeasonOnly] = useState<boolean>(false);
 	const [filtersExpanded, setFiltersExpanded] = useState<boolean>(true);
-	const [processingStatusFilters, setProcessingStatusFilters] = useState<ProcessingStatus[]>(
-		DEFAULT_PROCESSING_STATUS_FILTERS
-	);
-	const [processingStageFilter, setProcessingStageFilter] = useState<string | undefined>(undefined);
-	const [processingUserFilter, setProcessingUserFilter] = useState<string | undefined>(undefined);
-	const [selectedProcessingDatasetId, setSelectedProcessingDatasetId] = useState<number | null>(null);
-	const [logLimit, setLogLimit] = useState<number>(100);
-
 	// Update URL when filters change
 	useEffect(() => {
 		const params = new URLSearchParams();
@@ -301,55 +285,11 @@ function DatasetAuditInner() {
 		return Array.from(contributors).sort() as string[];
 	}, [contributorMap]);
 
-	const processingStageOptions = useMemo(() => {
-		const stages = new Set(
-			processingRows
-				.map((row) => row.current_status)
-				.filter((status): status is string => typeof status === "string" && status.length > 0)
-		);
-		return Array.from(stages)
-			.sort()
-			.map((value) => ({ label: value, value }));
-	}, [processingRows]);
-
-	const processingUserOptions = useMemo(() => {
-		const users = new Set(
-			processingRows
-				.map((row) => row.user_email)
-				.filter((email): email is string => typeof email === "string" && email.length > 0)
-		);
-		return Array.from(users)
-			.sort()
-			.map((value) => ({ label: value, value }));
-	}, [processingRows]);
-
-	const filteredProcessingRows = useMemo(() => {
-		return processingRows.filter((row) => {
-			if (processingStatusFilters.length > 0) {
-				const rowStatus = row.processing_status || "";
-				if (!processingStatusFilters.includes(rowStatus as ProcessingStatus)) return false;
-			}
-			if (processingStageFilter && row.current_status !== processingStageFilter) return false;
-			if (processingUserFilter && row.user_email !== processingUserFilter) return false;
-			return true;
-		});
-	}, [processingRows, processingStatusFilters, processingStageFilter, processingUserFilter]);
-
 	const processingCount = useMemo(() => {
 		return processingRows.filter((row) =>
-			DEFAULT_PROCESSING_STATUS_FILTERS.includes((row.processing_status || "") as ProcessingStatus)
+			PROCESSING_ACTIVE_STATUSES.some((status) => status === row.processing_status)
 		).length;
 	}, [processingRows]);
-
-	const selectedProcessingRow = useMemo(
-		() => processingRows.find((row) => row.dataset_id === selectedProcessingDatasetId) || null,
-		[processingRows, selectedProcessingDatasetId]
-	);
-
-	const { data: selectedDatasetLogs = [], isLoading: isLogsLoading, refetch: refetchSelectedDatasetLogs } = useDatasetLogs(
-		selectedProcessingDatasetId,
-		logLimit
-	);
 
 	// Filter datasets based on tab and filters
 	const filteredDatasets = useMemo(() => {
@@ -608,108 +548,6 @@ function DatasetAuditInner() {
 			/>
 		);
 	}
-
-	const processingColumns: ColumnsType<ProcessingOverviewRow> = [
-		{
-			title: "ID",
-			dataIndex: "dataset_id",
-			key: "dataset_id",
-			width: 90,
-			defaultSortOrder: "descend",
-			sorter: (a, b) => a.dataset_id - b.dataset_id,
-		},
-		{
-			title: "File",
-			dataIndex: "file_name",
-			key: "file_name",
-			width: 220,
-			ellipsis: true,
-			render: (value: string | null) => value || "—",
-		},
-		{
-			title: "Processing",
-			key: "processing_status",
-			width: 170,
-			render: (_, record) => {
-				const status = record.processing_status || "UNKNOWN";
-				const color =
-					status === "PROCESSING"
-						? "blue"
-						: status === "QUEUED"
-							? "orange"
-							: status === "FAILED"
-								? "red"
-								: "green";
-				return <Tag color={color} className="whitespace-nowrap">{status}</Tag>;
-			},
-		},
-		{
-			title: "Stage",
-			dataIndex: "current_status",
-			key: "current_status",
-			width: 130,
-			render: (value: string | null) => <Tag>{value || "idle"}</Tag>,
-		},
-		{
-			title: "Stuck",
-			key: "hours_in_current_status",
-			width: 100,
-			render: (_, record) => <span className="font-mono text-xs">{formatHours(record.hours_in_current_status)}</span>,
-			sorter: (a, b) => (a.hours_in_current_status || 0) - (b.hours_in_current_status || 0),
-		},
-		{
-			title: "Owner",
-			dataIndex: "user_email",
-			key: "user_email",
-			width: 220,
-			render: (value: string | null) => value || "—",
-		},
-		{
-			title: "Queue",
-			key: "queue",
-			width: 90,
-			render: (_, record) => (record.queue_priority === null ? "—" : `P${record.queue_priority}`),
-		},
-		{
-			title: "Error",
-			key: "error",
-			width: 200,
-			render: (_, record) => {
-				if (!record.has_error) return <span className="text-gray-400">—</span>;
-				const messageText = record.error_message || "Error";
-				const shortText = messageText.length > 80 ? `${messageText.slice(0, 80)}...` : messageText;
-				return (
-					<Tooltip title={messageText}>
-						<span className="text-red-600">{shortText}</span>
-					</Tooltip>
-				);
-			},
-		},
-		{
-			title: "Logs Preview",
-			key: "logs_preview",
-			width: 280,
-			render: (_, record) => {
-				if (!record.last_20_logs) return <span className="text-gray-400">No logs</span>;
-				const preview = record.last_20_logs.split("\n").slice(0, 2).join("\n");
-				return (
-					<Tooltip title={record.last_20_logs}>
-						<pre className="m-0 max-h-16 overflow-hidden whitespace-pre-wrap text-xs leading-tight">{preview}</pre>
-					</Tooltip>
-				);
-			},
-		},
-		{
-			title: "Actions",
-			key: "actions",
-			width: 120,
-			render: (_, record) => (
-				<Button size="small" onClick={() => setSelectedProcessingDatasetId(record.dataset_id)}>
-					View Logs
-				</Button>
-			),
-		},
-	];
 
 	// Build columns based on active tab
 	const baseColumns: ColumnsType<AuditDataset> = [
@@ -1180,169 +1018,14 @@ function DatasetAuditInner() {
 				</div>
 
 				<div className="rounded-2xl border border-gray-200/60 bg-white p-6 shadow-sm">
-					{activeTab === "processing" &&
-						(isProcessingError && processingRows.length === 0 ? (
-							<Result
-								status="error"
-								title="Could not load processing data"
-								subTitle="The processing overview could not be loaded. Please try again."
-								extra={<Button onClick={() => refetchProcessingRows()}>Try again</Button>}
-							/>
-						) : (
-							<>
-					{isProcessingError && (
-						<Alert
-							type="warning"
-							showIcon
-							message="Processing data may be stale"
-							description="The latest refresh failed. Previously loaded processing data remains visible."
-							action={<Button onClick={() => refetchProcessingRows()}>Try again</Button>}
-							className="mb-4"
+					{activeTab === "processing" && (
+						<ProcessingAuditTab
+							rows={processingRows}
+							isLoading={isProcessingLoading}
+							isError={isProcessingError}
+							onRefresh={() => void refetchProcessingRows()}
 						/>
 					)}
-					<div className="mb-4 flex flex-wrap items-end gap-3 rounded-md border border-gray-100 bg-gray-50 p-3">
-						<div>
-							<Text type="secondary" className="block mb-1 text-xs">
-								Processing Status
-							</Text>
-							<Select
-								mode="multiple"
-								allowClear
-								placeholder="Select statuses"
-								style={{ width: isMobile ? 280 : 360 }}
-								value={processingStatusFilters}
-								onChange={(value) => setProcessingStatusFilters(value as ProcessingStatus[])}
-								options={[
-									{ label: "Processing", value: "PROCESSING" },
-									{ label: "Queued", value: "QUEUED" },
-									{ label: "Failed", value: "FAILED" },
-									{ label: "Completed", value: "COMPLETED" },
-								]}
-							/>
-						</div>
-						<div>
-							<Text type="secondary" className="block mb-1 text-xs">
-								Stage
-							</Text>
-							<Select
-								allowClear
-								placeholder="All stages"
-								style={{ width: isMobile ? 160 : 200 }}
-								value={processingStageFilter}
-								onChange={setProcessingStageFilter}
-								options={processingStageOptions}
-								showSearch
-							/>
-						</div>
-						<div>
-							<Text type="secondary" className="block mb-1 text-xs">
-								Owner
-							</Text>
-							<Select
-								allowClear
-								placeholder="All owners"
-								style={{ width: isMobile ? 220 : 260 }}
-								value={processingUserFilter}
-								onChange={setProcessingUserFilter}
-								options={processingUserOptions}
-								showSearch
-							/>
-						</div>
-						<Space>
-							<Button onClick={() => refetchProcessingRows()}>Refresh</Button>
-							<Button
-								onClick={() => {
-									setProcessingStatusFilters(DEFAULT_PROCESSING_STATUS_FILTERS);
-									setProcessingStageFilter(undefined);
-									setProcessingUserFilter(undefined);
-								}}
-							>
-								Reset
-							</Button>
-						</Space>
-					</div>
-
-					<Table
-						dataSource={filteredProcessingRows}
-						columns={processingColumns}
-						rowKey="dataset_id"
-						loading={isProcessingLoading}
-						onRow={(record) => ({
-							onClick: () => setSelectedProcessingDatasetId(record.dataset_id),
-						})}
-						pagination={{
-							pageSize: 25,
-							showSizeChanger: true,
-							showQuickJumper: true,
-							showTotal: (total, range) => `${range[0]}-${range[1]} of ${total} datasets`,
-						}}
-						scroll={{ x: isMobile ? 980 : 1400 }}
-					/>
-
-					<Drawer
-						title={`Dataset ${selectedProcessingDatasetId || ""} Logs`}
-						open={selectedProcessingDatasetId !== null}
-						onClose={() => setSelectedProcessingDatasetId(null)}
-						width={isMobile ? "100%" : 820}
-					>
-						<div className="mb-3 flex flex-wrap items-center justify-between gap-2">
-							<Space>
-								<Text type="secondary">Log entries</Text>
-								<Select
-									value={logLimit}
-									onChange={setLogLimit}
-									style={{ width: 110 }}
-									options={[
-										{ label: "50", value: 50 },
-										{ label: "100", value: 100 },
-										{ label: "200", value: 200 },
-									]}
-								/>
-							</Space>
-							<Space>
-								{selectedProcessingRow?.current_status && <Tag>{selectedProcessingRow.current_status}</Tag>}
-								<Button onClick={() => refetchSelectedDatasetLogs()}>Refresh Logs</Button>
-							</Space>
-						</div>
-						<Table
-							rowKey="id"
-							size="small"
-							loading={isLogsLoading}
-							dataSource={selectedDatasetLogs}
-							pagination={false}
-							columns={[
-								{
-									title: "Time",
-									dataIndex: "created_at",
-									key: "created_at",
-									width: 170,
-									render: (value: string) => new Date(value).toLocaleString(),
-								},
-								{
-									title: "Level",
-									dataIndex: "level",
-									key: "level",
-									width: 90,
-									render: (value: string | null) => <Tag>{value || "INFO"}</Tag>,
-								},
-								{
-									title: "Category",
-									dataIndex: "category",
-									key: "category",
-									width: 140,
-									render: (value: string | null) => value || "general",
-								},
-								{
-									title: "Message",
-									dataIndex: "message",
-									key: "message",
-									render: (value: string | null) => <span className="whitespace-pre-wrap text-xs">{value || ""}</span>,
-								},
-							]}
-						/>
-					</Drawer>
-							</>
-						))}
 
 			{activeTab !== "processing" && (
 				<>
